// packages/lovdacn/src/preset/generated-theme-tokens.ts
export const THEME_TOKEN_NAMES = [
  "zinc",
  "slate",
  "stone",
  "gray",
  "neutral",
  "red",
  "orange",
  "amber",
  "yellow",
  "lime",
  "green",
  "emerald",
  "teal",
  "cyan",
  "sky",
  "blue",
  "indigo",
  "violet",
  "purple",
  "fuchsia",
  "pink",
  "rose",
  "mauve",
  "olive",
  "mist",
  "taupe"
] as const
export const BASE_COLOR_NAMES = [
  "zinc",
  "slate",
  "stone",
  "gray",
  "neutral",
  "taupe",
  "mauve",
  "olive",
  "mist"
] as const
// ...
export const THEME_TOKENS: Record<ThemeTokenName, ThemeTokenEntry> = {
  "zinc": {
    "kind": "base",
    "light": {
// ...
      "muted-foreground": "oklch(0.552 0.016 285.938)",
      "accent": "oklch(0.967 0.001 286.375)",
      "accent-foreground": "oklch(0.21 0.006 285.885)",
      "destructive": "oklch(0.577 0.245 27.325)",
      "border": "oklch(0.92 0.004 286.32)",
      "input": "oklch(0.92 0.004 286.32)",
      "ring": "oklch(0.705 0.015 286.067)",
      "chart-1": "oklch(0.871 0.006 286.286)",
      "chart-2": "oklch(0.552 0.016 285.938)",
      "chart-3": "oklch(0.442 0.017 285.786)",
      "chart-4": "oklch(0.37 0.013 285.805)",
      "chart-5": "oklch(0.274 0.006 286.033)",
      "sidebar": "oklch(0.985 0 0)",
      "sidebar-foreground": "oklch(0.141 0.005 285.823)",
      "sidebar-primary": "oklch(0.21 0.006 285.885)",
      "sidebar-primary-foreground": "oklch(0.985 0 0)",
      "sidebar-accent": "oklch(0.967 0.001 286.375)",
      "sidebar-accent-foreground": "oklch(0.21 0.006 285.885)",
      "sidebar-border": "oklch(0.92 0.004 286.32)",
      "sidebar-ring": "oklch(0.705 0.015 286.067)"
    },
// ...
  },
// ...
}